requirements.txt
gitdb==4.0.12
GitPython==3.1.46
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
platformdirs==4.9.2
Pygments==2.19.2
rich==14.3.2
smmap==5.0.2
textual==8.0.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
